Move AppDomainInfo
