+++ src/Microsoft.Diagnostics.Runtime/AbstractDac/IClrRuntimeHelpers.cs
@@ -49,6 +49,27 @@ namespace Microsoft.Diagnostics.Runtime.AbstractDac
         void Flush();
     }
 
+    internal struct AppDomainInfo
+    {
+        public ulong Address { get; set; }
+        public AppDomainKind Kind { get; set; }
+
+        public int Id { get; set; }
+        public string Name { get; set; }
+
+        public string? ConfigFile { get; set; }
+        public string? ApplicationBase { get; set; }
+
+        public ulong LoaderAllocator { get; set; }
+    }
+
+    internal enum AppDomainKind
+    {
+        Normal,
+        System,
+        Shared,
+    }
+
     /// <summary>
     /// Information about a ClrHandle
     /// </summary>
